+++ examples/terraform/fedora-coreos-install/groups.tf
@@ -1,0 +1,6 @@
+// Default matcher group for machines
+resource "matchbox_group" "default" {
+  name    = "default"
+  profile = matchbox_profile.fedora-coreos-install.name
+}
+