+++ CHANGES.md
@@ -8,7 +8,7 @@ Notable changes between releases.
   * Publish Matchbox images from internal infra to Quay (`quay.io/poseidon/matchbox`)
   * Update Go version from v1.13.4 to v1.14.9
 * Update base image from `alpine:3.10` to `alpine:3.12` ([#784](https://github.com/poseidon/matchbox/pull/784))
-* Refresh docs and examples for Fedora CoreOS and Flatcar Linux ([#815](https://github.com/poseidon/matchbox/pull/815))
+* Refresh docs and examples for Fedora CoreOS and Flatcar Linux ([#815](https://github.com/poseidon/matchbox/pull/815), [#816](https://github.com/poseidon/matchbox/pull/816))
 * Update Kubernetes manifest examples ([#791](https://github.com/poseidon/matchbox/pull/791))
 * Include `contrib/k8s` in release tarballs ([#788](https://github.com/poseidon/matchbox/pull/788))
 * Remove RPM spec file (Copr publishing stopped in v0.6)